Final back 2

+++ backend/src/controllers/class.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import Class from "../models/class";
+import { getUserRole } from "./user.controller";
 
 export const getAllClasses = async (
   req: Request,
@@ -21,6 +22,11 @@ export const createClass = async (
   req: Request,
   res: Response
 ): Promise<Response> => {
+  const role = getUserRole(req);
+  if (role !== "admin" && role !== "teacher") {
+    return res.status(401).json({ error: "Unauthorized" });
+  }
+
   const { language, teacher, students, schedule } = req.body;
   const newClass = new Class({ language, teacher, students, schedule });
   await newClass.save();
@@ -31,6 +37,11 @@ export const updateClass = async (
   req: Request,
   res: Response
 ): Promise<Response> => {
+  const role = getUserRole(req);
+  if (role !== "admin" && role !== "teacher") {
+    return res.status(401).json({ error: "Unauthorized" });
+  }
+
   const { id } = req.params;
   const updatedClass = await Class.findByIdAndUpdate(id, req.body, {
     new: true,
@@ -42,6 +53,11 @@ export const deleteClass = async (
   req: Request,
   res: Response
 ): Promise<Response> => {
+  const role = getUserRole(req);
+  if (role !== "admin" && role !== "teacher") {
+    return res.status(401).json({ error: "Unauthorized" });
+  }
+
   await Class.findByIdAndDelete(req.params.id);
   return res.json({
     msg: "Class Deleted Successfully",
